// src/app/features/casino/models/poker-data.model.ts
export interface PokerData {
    nothingMultiplicator: number;
    onePairMultiplicator: number;
    twoPairMultiplicator: number;
    threeOfAKindMultiplicator: number;
    straightMultiplicator: number;
    flushMultiplicator: number;
    fullHouseMultiplicator: number;
    fourOfAKindMultiplicator: number;
    straightFlushMultiplicator: number;
    royalFlushMultiplicator: number;
    maxBet: number;
    nothingVIPMultiplicator: number;
    onePairVIPMultiplicator: number;
    twoPairVIPMultiplicator: number;
    threeOfAKindVIPMultiplicator: number;
    straightVIPMultiplicator: number;
    flushVIPMultiplicator: number;
    fullHouseVIPMultiplicator: number;
    fourOfAKindVIPMultiplicator: number;
    straightFlushVIPMultiplicator: number;
    royalFlushVIPMultiplicator: number;
    maxVIPBet: number;
}